Move start.py to automation module inside
