+++ bdfrtohtml/__main__.py
@@ -9,14 +9,14 @@ import click
 import shutil
 from bdfrtohtml import filehelper
 from bdfrtohtml import posthelper
+from bdfrtohtml import automation
 from bdfrtohtml import util
 import logging
 import copy
 
 logger = logging.getLogger(__name__)
 
-
-@click.command()
+@click.group(invoke_without_command=True)
 @click.option('--input_folder', default=None, help='The folder where the download and archive results have been saved to')
 @click.option('--output_folder', default=None, help='Folder where the HTML results should be created.')
 @click.option('--recover_comments', type=bool, help='Should we attempt to recover deleted comments?')
@@ -32,6 +32,8 @@ logger = logging.getLogger(__name__)
 @click.pass_context
 def main(context: click.Context, **_):
 
+    if context.invoked_subcommand is not None:
+        return
     if context.params.get('config'):
         config = util.load_config(context.params.get('config'))
     else:
@@ -89,6 +91,10 @@ def main(context: click.Context, **_):
 
 
     logging.info("BDFR-HTML run complete.")
+
+@main.command("automate")
+def run_automation():
+    automation.automate()
 
 
 if __name__ == '__main__':
